Inventory System's Done (for now)
All new items to be picked up need to be a child of the BP_PickupBase with a new Construct Class node of BP_InventoryItem in the Construction Script. Future-proofed logic to make certain "keys" unlock doors has not been added yet. However, more item data can be added to BP_InventoryItem if needed. Any interactable world objects will likely need to be made separately but under the same Interface as the inventory items, BPI_InteractableObject.