Door and Key Functionality
If a door has a key attached to it, then that key is needed to open the door. If not, then the door can open as usual. Key's should be able to work across levels, as long as the door it's used for has an EXACT copy of the key in an inaccessible area of the map and we have an inventory save and load system.